# apis/api-specs.yml
openapi: 3.0.3
info:
  title: Uploader API
  version: 1.3.0
  description: |
    OpenAPI specification for the uploader service.

    The service accepts multipart file uploads, stores file metadata,
    and returns a short-lived signed download redirect.

    User-scoped endpoints (/v1/user/*) require authentication via oauth2-proxy.
    The proxy validates user sessions and injects X-Auth-Email into upstream
    requests sent to this service.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
  /v1/files:
    post:
      tags:
        - Files
      summary: Upload a file
      description: |
        Streams a multipart file upload using the `file` form field.

        A password may be supplied in the `password` multipart form field.
        When supplied, it is stored as a hash and must be presented again to
        download that protected file.

        An optional expiration timestamp may be supplied in the
        `expires_at` multipart form field using RFC3339 format (UTC recommended).
        After expiration, downloads are denied and the system may remove the
        object and metadata during cleanup.

        The service enforces a configurable maximum upload size.
        By default, uploads larger than 50 MB are rejected.
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: File content to upload.
                password:
                  type: string
                  minLength: 1
                  description: Optional password used to protect downloads for the uploaded file.
                expires_at:
                  type: string
                  format: date-time
                  description: Optional RFC3339 timestamp after which the file is no longer downloadable.
      responses:
        '201':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'
              examples:
                uploaded:
                  value:
                    file_id: 550e8400-e29b-41d4-a716-446655440000
                    filename: file.pdf
                    content_type: application/pdf
                    size_bytes: 12345
        '400':
          description: Invalid multipart form data or missing file metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidMultipart:
                  value:
                    error: invalid multipart form
                missingFile:
                  value:
                    error: missing file field
                missingFilename:
                  value:
                    error: file name is required
                invalidExpiration:
                  value:
                    error: invalid expiration; use RFC3339 timestamp
                pastExpiration:
                  value:
                    error: expiration must be in the future
        '413':
          description: Uploaded file exceeds the configured maximum size
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: file exceeds max size
        '500':
          description: Upload or metadata persistence failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                uploadFailed:
                  value:
                    error: upload failed
                persistFailed:
                  value:
                    error: failed to persist metadata
                secureFileFailed:
                  value:
                    error: failed to secure file
    get:
      tags:
        - Files
      summary: List files
      description: |
        Returns a paginated list of uploaded files ordered by creation date (newest first).

        Supports pagination using `limit` and `offset` query parameters.
        Default limit is 50, maximum limit is 1000.
        Results are ordered by creation date in descending order.
      operationId: listFiles
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum number of files to return. Default is 50, maximum is 1000.
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 1000
        - name: offset
          in: query
          required: false
          description: Number of files to skip for pagination. Default is 0.
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Files retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileListResponse'
              examples:
                empty:
                  value:
                    files: []
                    limit: 50
                    offset: 0
                withFiles:
                  value:
                    files:
                      - id: 550e8400-e29b-41d4-a716-446655440000
                        filename: document.pdf
                        content_type: application/pdf
                        size_bytes: 24576
                        expires_at: '2026-05-21T10:30:00Z'
                        created_at: '2026-05-20T10:00:00Z'
                      - id: 550e8400-e29b-41d4-a716-446655440001
                        filename: image.png
                        content_type: image/png
                        size_bytes: 102400
                        created_at: '2026-05-20T09:30:00Z'
                    limit: 50
                    offset: 0
        '500':
          description: Failed to retrieve file list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: failed to list files
  /v1/files/{file_id}/download:
    get:
      tags:
        - Files
      summary: Get a signed download redirect
      description: |
        Looks up the uploaded file by identifier and returns an HTTP `302`
        redirect to a signed object storage URL.

        This endpoint works for files without a password.
        For password-protected files, use the `POST` variant of this endpoint
        and send the password in the request body.

        The signed URL currently has a 60-second time-to-live.
      operationId: downloadFileGet
      parameters:
        - name: file_id
          in: path
          required: true
          description: Unique file identifier returned by the upload endpoint.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to a signed download URL
          headers:
            Location:
              description: Signed object storage URL for downloading the file.
              schema:
                type: string
                format: uri
        '400':
          description: Missing file identifier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: file_id is required
        '401':
          description: Password required for a protected file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: password is required
        '404':
          description: File metadata not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: file not found
        '500':
          description: Metadata lookup or URL signing failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                metadataFetchFailed:
                  value:
                    error: failed to fetch metadata
                signFailed:
                  value:
                    error: failed to sign url
    post:
      tags:
        - Files
      summary: Get a signed download redirect for a protected file
      description: |
        Looks up the uploaded file by identifier and returns an HTTP `302`
        redirect to a signed object storage URL.

        For password-protected files, clients must send the same password used
        during upload in the request body.
        For files uploaded without a password, no password is required.

        The signed URL currently has a 60-second time-to-live.
      operationId: downloadFilePost
      parameters:
        - name: file_id
          in: path
          required: true
          description: Unique file identifier returned by the upload endpoint.
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadPasswordRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DownloadPasswordRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DownloadPasswordRequest'
      responses:
        '302':
          description: Redirect to a signed download URL
          headers:
            Location:
              description: Signed object storage URL for downloading the file.
              schema:
                type: string
                format: uri
        '400':
          description: Invalid request body or missing file identifier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingFileID:
                  value:
                    error: file_id is required
                invalidBody:
                  value:
                    error: invalid download request body
        '401':
          description: Password missing or invalid for a protected file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingPassword:
                  value:
                    error: password is required
                invalidPassword:
                  value:
                    error: invalid password
        '404':
          description: File metadata not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: file not found
        '500':
          description: Metadata lookup or URL signing failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                metadataFetchFailed:
                  value:
                    error: failed to fetch metadata
                signFailed:
                  value:
                    error: failed to sign url
  /v1/user/files:
    get:
      tags:
        - User
      summary: List authenticated user's files
      description: |
        Returns a paginated list of files owned by the authenticated user.

        Identity is resolved from the X-Auth-Email header injected by
        oauth2-proxy. Guest uploads (without owner identity) are excluded.
      operationId: listUserFiles
      security:
        - oauth2Proxy: []
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum number of files to return. Default 50, maximum 100.
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          description: Number of files to skip for pagination. Default 0.
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Files retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserFileListResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: authentication required
        '500':
          description: Failed to retrieve file list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: failed to list files
  /v1/user/files/{file_id}:
    delete:
      tags:
        - User
      summary: Delete an owned file
      description: |
        Deletes a file only when the authenticated user's identity matches
        the file owner.

        Returns 404 both for missing files and files belonging to other users
        to reduce information disclosure risks.
      operationId: deleteUserFile
      security:
        - oauth2Proxy: []
      parameters:
        - name: file_id
          in: path
          required: true
          description: Unique file identifier returned by the upload endpoint.
          schema:
            type: string
      responses:
        '204':
          description: File deleted successfully
        '400':
          description: Missing file identifier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: file_id is required
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: authentication required
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: file not found
        '500':
          description: File deletion failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                storageDeleteFailed:
                  value:
                    error: failed to delete file
                metadataDeleteFailed:
                  value:
                    error: failed to delete file metadata
  /v1/user/files/{file_id}/stats:
    get:
      tags:
        - User
      summary: Get stats for an owned file
      description: |
        Returns metadata and download statistics for a single owned file.

        Returns 404 for files that do not exist or are owned by another user.
      operationId: getUserFileStats
      security:
        - oauth2Proxy: []
      parameters:
        - name: file_id
          in: path
          required: true
          description: Unique file identifier returned by the upload endpoint.
          schema:
            type: string
      responses:
        '200':
          description: File stats retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserFileView'
        '400':
          description: Missing file identifier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: file_id is required
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: authentication required
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: file not found
        '500':
          description: Metadata lookup failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: failed to fetch file
...
components:
  schemas:
    DownloadPasswordRequest:
      type: object
      properties:
        password:
          type: string
          minLength: 1
          description: Password required to authorize download of a protected file.
    FileUploadResponse:
      type: object
      required:
        - file_id
        - filename
        - content_type
        - size_bytes
      properties:
        file_id:
          type: string
          description: Server-generated file identifier.
          example: 550e8400-e29b-41d4-a716-446655440000
        filename:
          type: string
          description: Sanitized original file name.
          example: file.pdf
        content_type:
          type: string
          description: Media type associated with the uploaded file.
          example: application/pdf
        size_bytes:
          type: integer
          format: int64
          description: Number of bytes streamed for the uploaded file.
          example: 12345
        expires_at:
          type: string
          format: date-time
          description: Optional RFC3339 expiration timestamp when provided at upload time.
          example: 2026-05-07T16:30:00Z
...
    UserFileListResponse:
      type: object
      required:
        - files
        - limit
        - offset
      properties:
        files:
          type: array
          description: List of owned file records.
          items:
            $ref: '#/components/schemas/UserFileView'
        limit:
          type: integer
          description: The limit parameter used in the request.
          example: 50
        offset:
          type: integer
          description: The offset parameter used in the request.
          example: 0
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message.
          example: invalid multipart form
  securitySchemes:
    oauth2Proxy:
      type: apiKey
      in: header
      name: X-Auth-Email
      description: |
        Identity header injected by oauth2-proxy after successful user
        authentication. This header is trusted only when requests are routed
        through the proxy.